Factored everything but showcase component
